dev: support custom serializer
