+++ app/src/main/AndroidManifest.xml
@@ -18,6 +18,19 @@
                 <category android:name="android.intent.category.LAUNCHER" />
             </intent-filter>
         </activity>
+
+        <provider
+            android:name="androidx.startup.InitializationProvider"
+            android:authorities="${applicationId}.androidx-startup"
+            android:exported="false"
+            tools:node="merge">
+<!--            <meta-data-->
+<!--                android:name="io.wax911.emojifysample.CustomEmojiInitializer"-->
+<!--                android:value="androidx.startup" />-->
+            <meta-data
+                android:name="io.wax911.emojify.initializer.EmojiInitializer"
+                android:value="androidx.startup" />
+        </provider>
     </application>
 
 </manifest>
