--- emojify/src/main/AndroidManifest.xml
@@ -1,15 +1,0 @@
-<manifest xmlns:android="http://schemas.android.com/apk/res/android"
-    xmlns:tools="http://schemas.android.com/tools">
-
-    <application>
-        <provider
-            android:name="androidx.startup.InitializationProvider"
-            android:authorities="${applicationId}.androidx-startup"
-            android:exported="false"
-            tools:node="merge">
-            <meta-data
-                android:name="io.wax911.emojify.initializer.EmojiInitializer"
-                android:value="androidx.startup" />
-        </provider>
-    </application>
-</manifest>
