+++ gradle/libs.versions.toml
@@ -21,6 +21,7 @@ jetbrains-kotlinx-datetime = "0.5.0"
 jetbrains-kotlinx-serialization = "1.6.2"
 
 mockk = "1.13.9"
+moshi-kotlin = "1.15.0"
 
 
 [plugins]
@@ -28,6 +29,7 @@ gradle-versions = { id = "com.github.ben-manes.versions", version = "0.51.0" }
 
 
 [libraries]
+moshi-kotlin = { module = "com.squareup.moshi:moshi-kotlin", version.ref = "moshi-kotlin" }
 timber = { module = "com.jakewharton.timber:timber", version = "5.0.1" }
 junit = { module = "junit:junit", version = "4.13.2" }
