+++ gradle/version.properties
@@ -1,2 +1,2 @@
-version=1.7.1
-code=1007001000
+version=2.7.1
+code=2007001000
